requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.34.2
httpx==0.27.0
pymupdf==1.24.10
python-dotenv==1.0.1
pillow==10.4.0
python-multipart==0.0.12
